extract login/signup form from FormDialog to exclusive component
